What's New BETA (Photon OS 5.0)
[Squashed]
The following commits are squashed:

- This commit consists of the final Changes before BETA release.

- What's New in Photon OS 5.0 Draft 1

This is the first draft of Photon OS 5.0 What's New section. It lists all the BETA features for now.

- What's New in Photon OS 5.0 Draft 2 (Review Comments Incorporated)

This is the second draft of Photon OS 5.0 What's New. All the review comments received before 26th January are incorporated in this commit.

- Adding Network Event Broker on What's New

Adding a one-liner for Network Event Broker on What's New.